Add clearsilver, a templating language.
This build is linked against our python 2.6, which is 64bit. This means
the package is 64bit-only. There are also some patches to the build
system to make it work on non-32bit GNU/Linux. Upstream hasn't released
in about three years, so I'm doubtful these patches will be merged.
I'll make the effort, though.