Update and re-build ruby against our new gcc.
This has required a few things to change to keep the world turning.
There's an issue with rubygems using mkmf (which only generates GNU
make-compatible Makefiles) and the ffi gem (among others) will not build
otherwise. This is fixed in a pull request accepted into rubygems
(https://github.com/rubygems/rubygems/pull/443). As a side-effect, our
ruby packages now need to (rather, always should have, but it was never
respected) specify MAKE=gmake, which I've updated for the other ports
included in this commit.

vagrant has some issues related to gem activation separate from this
change (they manifest in the vagrant package in the ms repo today), so
I've not modified that package.

This addresses #19 and brings us up to date with the ruby 1.9 branch,
patching a number of bugs and security issues along the way as well.
Special thanks to @erikh and @raggi without whom this work wouldn't
have been possible.