Trim extraneous text from OmniTI CA pem file.
When pkg(1) sets a CA cert as approved, it remembers the hash
of the bare certificate block, not the hash of the actual file
that contained the certificate block.  If the origin file contains
additional text, making its manifest checksum different, pkg(1)
will be unable to find it in the repo.

Refs #58