Fix catastrophic typo. :)
A typo changed the name of our local make_install function,
causing us to run a full make install of all of Postgres, rather
than just in the contrib/citext directory. This resulted in a
citext package that contained the entire postgres install and
conflicted with the main postgresql-92x package.