Add a newline to headers.sh for ncurses.
RARELY, but it does happen, a newline isn't inserted between two lines of
commands for headers.sh.  This causes the build to die.  There is likely a
root cause for this occasional failure, but given it's proximit to r151018's
release date, I choose instead this simple hack. An extra newline between
command in a shell script never hurts, but in this case, not having any hurts
a lot.